fix: Move backdated entry validation from transaction base to stock ledger entry
