+++ erpnext/stock/doctype/stock_ledger_entry/stock_ledger_entry.py
@@ -5,7 +5,7 @@
 from __future__ import unicode_literals
 import frappe
 from frappe import _
-from frappe.utils import flt, getdate, add_days, formatdate
+from frappe.utils import flt, getdate, add_days, formatdate, get_datetime, date_diff
 from frappe.model.document import Document
 from datetime import date
 from erpnext.controllers.item_variant import ItemTemplateCannotHaveStock
@@ -33,6 +33,8 @@ class StockLedgerEntry(Document):
 		self.scrub_posting_time()
 		self.validate_and_set_fiscal_year()
 		self.block_transactions_against_group_warehouse()
+		self.validate_with_last_transaction_posting_time()
+		self.validate_future_posting()
 
 	def on_submit(self):
 		self.check_stock_frozen_date()
@@ -138,6 +140,30 @@ class StockLedgerEntry(Document):
 	def block_transactions_against_group_warehouse(self):
 		from erpnext.stock.utils import is_group_warehouse
 		is_group_warehouse(self.warehouse)
+
+	def validate_with_last_transaction_posting_time(self):
+		last_transaction_time = frappe.db.sql("""
+			select MAX(timestamp(posting_date, posting_time)) as posting_time
+			from `tabStock Ledger Entry`
+			where docstatus = 1 and item_code = %s
+			and warehouse = %s""", (self.item_code, self.warehouse))[0][0]
+
+		cur_doc_posting_datetime = "%s %s" % (self.posting_date, self.get("posting_time") or "00:00:00")
+
+		if last_transaction_time and get_datetime(cur_doc_posting_datetime) < get_datetime(last_transaction_time):
+			msg = _("Last Stock Transaction for item {0} under warehouse {1} was on {2}.").format(frappe.bold(self.item_code),
+				frappe.bold(self.warehouse), frappe.bold(last_transaction_time))
+
+			msg += "<br><br>" + _("Stock Transactions for Item {0} under warehouse {1} cannot be posted before this time.").format(
+				frappe.bold(self.item_code), frappe.bold(self.warehouse))
+
+			msg += "<br><br>" + _("Please remove this item and try to submit again or update the posting time.")
+			frappe.throw(msg, title=_("Backdated Stock Entry"))
+
+	def validate_future_posting(self):
+		if date_diff(self.posting_date, getdate()) > 0:
+			msg = _("Posting future stock transactions are not allowed due to Immutable Ledger")
+			frappe.throw(msg, title=_("Future Posting Not Allowed"))
 
 def on_doctype_update():
 	if not frappe.db.has_index('tabStock Ledger Entry', 'posting_sort_index'):
