+++ erpnext/utilities/transaction_base.py
@@ -29,28 +29,6 @@ class TransactionBase(StatusUpdater):
 			except ValueError:
 				frappe.throw(_('Invalid Posting Time'))
 
-		self.validate_future_posting()
-		self.validate_with_last_transaction_posting_time()
-
-	def is_stock_transaction(self):
-		if self.doctype not in ["Sales Invoice", "Purchase Invoice", "Stock Entry", "Stock Reconciliation",
-			"Delivery Note", "Purchase Receipt", "Fees"]:
-				return False
-
-		if self.doctype in ["Sales Invoice", "Purchase Invoice"]:
-			if not (self.get("update_stock") or self.get("is_pos")):
-				return False
-
-		return True
-
-	def validate_future_posting(self):
-		if not self.is_stock_transaction():
-			return
-
-		if getattr(self, 'set_posting_time', None) and date_diff(self.posting_date, nowdate()) > 0:
-			msg = _("Posting future transactions are not allowed due to Immutable Ledger")
-			frappe.throw(msg, title=_("Future Posting Not Allowed"))
-
 	def add_calendar_event(self, opts, force=False):
 		if cstr(self.contact_by) != cstr(self._prev.contact_by) or \
 				cstr(self.contact_date) != cstr(self._prev.contact_date) or force or \
@@ -179,26 +157,6 @@ class TransactionBase(StatusUpdater):
 			ret = None
 
 		return ret
-
-	def validate_with_last_transaction_posting_time(self):
-
-		if not self.is_stock_transaction():
-			return
-
-		for item in self.get('items'):
-			last_transaction_time = frappe.db.sql("""
-				select MAX(timestamp(posting_date, posting_time)) as posting_time
-				from `tabStock Ledger Entry`
-				where docstatus = 1 and item_code = %s
-				and company = %s""", (item.item_code, self.get('company')))[0][0]
-
-			cur_doc_posting_datetime = "%s %s" % (self.posting_date, self.get("posting_time") or "00:00:00")
-
-			if last_transaction_time and get_datetime(cur_doc_posting_datetime) < get_datetime(last_transaction_time):
-				msg = _("Last Stock Transaction for item {0} was on {1}.").format(frappe.bold(item.item_code), frappe.bold(last_transaction_time))
-				msg += "<br><br>" + _("Stock Transactions for Item {0} cannot be posted before this time.").format(frappe.bold(item.item_code))
-				msg += "<br><br>" + _("Please remove this item and try to submit again or update the posting time.")
-				frappe.throw(msg, title=_("Backdated Stock Entry"))
 
 def delete_events(ref_type, ref_name):
 	events = frappe.db.sql_list(""" SELECT
